Rename features -> langaugeFeatures
