+++ extensions/typescript-language-features/src/languageFeatures/bufferSyncSupport.ts
