+++ extensions/typescript-language-features/src/languageFeatures/completions.ts
