+++ extensions/typescript-language-features/src/languageFeatures/directiveCommentCompletions.ts
