+++ extensions/typescript-language-features/src/languageFeatures/documentHighlight.ts
