+++ extensions/typescript-language-features/src/languageFeatures/documentSymbol.ts
