+++ extensions/typescript-language-features/src/languageFeatures/fileConfigurationManager.ts
