+++ extensions/typescript-language-features/src/languageFeatures/folding.ts
