+++ extensions/typescript-language-features/src/languageFeatures/rename.ts
