+++ extensions/typescript-language-features/src/languageFeatures/smartSelect.ts
