+++ extensions/typescript-language-features/src/languageFeatures/tagClosing.ts
