+++ extensions/typescript-language-features/src/languageFeatures/updatePathsOnRename.ts
