+++ extensions/typescript-language-features/src/languageFeatures/workspaceSymbols.ts
